rename version module to _version
